First implementation of repair mechanism
